added materialize-helmet to packages within client folder renamed pages component to materializePage, which links in materialize css, js, and google fonts, so any materialize used within this wrapper will work tested out materialize Navbar in the footer, it works as long as the footer component is within a Materialize Wrapper
