+++ client/package.json
@@ -8,6 +8,7 @@
     "chart.js": "^2.7.2",
     "react": "^16.1.1",
     "react-dom": "^16.1.1",
+    "react-helmet": "^5.2.0",
     "react-materialize": "^2.3.3",
     "react-router-dom": "^4.2.2",
     "react-scripts": "^1.0.17"
